Split Symbol Column & Fix Total Column
Add a new `getExchangeInfo` API function for fetching the details of
multiple symbols. This lets us extract the `symbol` column in the export
into separate `base-asset` & `quote-asset` columns. It also allows us to
use Binance's precision for each asset when rendering an export row.

Fix the `total` column in the export. The value returned by the Binance
API is truncated at the quote asset's precision, but the true value may
require higher precision. To fix this, we ignore the `tQuoteQuantity`
from the API & calculate it ourselves by multiplying the `tPrice` with
the `tAmount`.

Move the CSV serialization to the `Main` module, add a helper type since
serialization now requires both the `SymbolDetails` along with the
`Trade`, & add a helper `BinanceApiM` type to simplify some type
signatures.